+++ src/app/units/states/edit/directives/unit-staff-editor/unit-staff-editor.component.ts
@@ -8,6 +8,7 @@ import {MatTableDataSource} from '@angular/material/table';
 import {MatButtonToggleChange} from '@angular/material/button-toggle';
 import {ConfirmationModalService} from 'src/app/common/modals/confirmation-modal/confirmation-modal.service';
 import {MatSelectChange} from '@angular/material/select';
+import {TutorNotesModalService} from 'src/app/common/modals/tutor-notes-modal/tutor-notes-modal.service';
 
 @Component({
   selector: 'unit-staff-editor',
@@ -39,6 +40,7 @@ export class UnitStaffEditorComponent implements OnInit {
     private alertService: AlertService,
     private unitRoleService: UnitRoleService,
     private confirmationModalService: ConfirmationModalService,
+    private tutorNotesModal: TutorNotesModalService,
   ) {}
 
   ngOnInit(): void {
@@ -208,13 +210,24 @@ export class UnitStaffEditorComponent implements OnInit {
    * @returns void
    */
   removeStaff(staff: UnitRole) {
-    this.unitRoleService.delete(staff, {cache: this.unit.staffCache}).subscribe({
-      next: () => this.alertService.success('Staff member removed', 2000),
-      error: (response) => this.alertService.error(response, 6000),
-    });
+    this.confirmationModalService.show(
+      'Remove staff member',
+      `Are you sure you want to remove ${staff.user.name} from ${this.unit.code} ${this.unit.name}?`,
+      () => {
+        this.unitRoleService.delete(staff, {cache: this.unit.staffCache}).subscribe({
+          next: () => this.alertService.success('Staff member removed', 2000),
+          error: (response) => this.alertService.error(response, 6000),
+        });
+      },
+    );
   }
 
   groupSetName(id: number) {
     this.unit.groupSetsCache.get(id).name || 'Individual Work';
+  }
+
+  openTutorNotes(unitRole: UnitRole) {
+    unitRole.unit = this.unit; // HACK: ensure unit is mapped within the UnitRole
+    this.tutorNotesModal.show(null, unitRole);
   }
 }